functionCaller restructure
