+++ src/utils/functionCaller.ts
@@ -1,10 +1,19 @@
-import axios from 'axios'
-import { useSettingsStore } from '../stores/settingsStore'
 import { createEmbedding } from '../services/apiService'
+import { manage_clipboard } from './functions/clipboard'
 import {
-  parseNaturalLanguageToCron,
-  validateCronExpression,
-} from './cronParser'
+  open_path,
+  list_directory,
+  execute_command,
+} from './functions/filesystem'
+import {
+  schedule_task,
+  manage_scheduled_tasks,
+  get_calendar_events,
+  create_calendar_event,
+  update_calendar_event,
+  delete_calendar_event,
+} from './functions/calendar'
+import { search_torrents, add_torrent_to_qb } from './functions/torrent'
 
 interface Crawl4AiArgs {
   url: string
@@ -14,18 +23,6 @@ interface FunctionResult {
   success: boolean
   data?: any
   error?: string
-}
-
-interface OpenPathArgs {
-  target: string
-}
-
-interface TorrentSearchArgs {
-  query: string
-}
-
-interface AddTorrentArgs {
-  magnet: string
 }
 
 interface SaveMemoryArgs {
@@ -53,26 +50,6 @@ interface SearchEmailsArgs {
 
 interface GetEmailContentArgs {
   messageId: string
-}
-
-interface ListDirectoryArgs {
-  path: string
-}
-
-interface ExecuteCommandArgs {
-  command: string
-}
-
-interface ScheduleTaskArgs {
-  name: string
-  schedule: string
-  action_type: 'command' | 'reminder'
-  details: string
-}
-
-interface ManageScheduledTasksArgs {
-  action: 'list' | 'delete' | 'toggle'
-  task_id?: string
 }
 
 interface BrowserContextArgs {
@@ -261,316 +238,6 @@ async function get_current_datetime(args: {
 }
 
 /**
- * Opens a file, folder, application, or URL using the system's default handler.
- */
-async function open_path(args: OpenPathArgs): Promise<FunctionResult> {
-  console.log(`Invoking open_path with target: ${args.target}`)
-
-  try {
-    if (typeof window === 'undefined' || !window.ipcRenderer?.invoke) {
-      return {
-        success: false,
-        error:
-          'Electron IPC bridge not available. This function only works in the desktop app.',
-      }
-    }
-
-    const result = await window.ipcRenderer.invoke('electron:open-path', args)
-    console.log('Main process response for open_path:', result)
-
-    if (result.success) {
-      return { success: true, data: { message: result.message } }
-    } else {
-      return { success: false, error: result.message }
-    }
-  } catch (error) {
-    console.error('Error invoking electron:open-path:', error)
-    return {
-      success: false,
-      error: `Failed to execute open_path: ${error instanceof Error ? error.message : 'Unknown error'}`,
-    }
-  }
-}
-
-/**
- * Manages the system clipboard by reading or writing text content.
- */
-async function manage_clipboard(args: {
-  action: 'read' | 'write'
-  content?: string
-}): Promise<FunctionResult> {
-  console.log(`Invoking clipboard action: ${args.action}`)
-
-  try {
-    if (typeof window === 'undefined' || !window.ipcRenderer?.invoke) {
-      return {
-        success: false,
-        error:
-          'Electron IPC bridge not available. This function only works in the desktop app.',
-      }
-    }
-
-    if (args.action !== 'read' && args.action !== 'write') {
-      return {
-        success: false,
-        error: 'Invalid clipboard action. Must be "read" or "write".',
-      }
-    }
-
-    if (
-      args.action === 'write' &&
-      (args.content === undefined || args.content === null)
-    ) {
-      return {
-        success: false,
-        error: 'Content is required for clipboard write operations.',
-      }
-    }
-
-    const result = await window.ipcRenderer.invoke(
-      'electron:manage-clipboard',
-      args
-    )
-    console.log('Main process response for clipboard operation:', result)
-
-    if (result.success) {
-      if (args.action === 'read' && result.data !== undefined) {
-        return {
-          success: true,
-          data: result.data,
-        }
-      }
-
-      return {
-        success: true,
-        data: { message: result.message },
-      }
-    } else {
-      return {
-        success: false,
-        error: result.message,
-      }
-    }
-  } catch (error) {
-    console.error('Error during clipboard operation:', error)
-    return {
-      success: false,
-      error: `Failed to perform clipboard operation: ${error instanceof Error ? error.message : 'Unknown error'}`,
-    }
-  }
-}
-
-/**
- * Searches for torrents using the Torrent Search API.
- */
-
-async function search_torrents(
-  args: TorrentSearchArgs
-): Promise<FunctionResult> {
-  const settings = useSettingsStore().config
-  const JACKETT_API_KEY = settings.VITE_JACKETT_API_KEY
-  const JACKETT_URL = settings.VITE_JACKETT_URL
-
-  if (!JACKETT_API_KEY || !JACKETT_URL || !args.query) {
-    return { success: false, error: 'Missing Jackett configuration or query.' }
-  }
-
-  try {
-    const url = `${JACKETT_URL}/api/v2.0/indexers/all/results/torznab/api?apikey=${JACKETT_API_KEY}&t=search&q=${encodeURIComponent(args.query)}`
-    const response = await axios.get(url, { responseType: 'text' })
-
-    const parser = new DOMParser()
-    const xml = parser.parseFromString(response.data, 'application/xml')
-    const items = Array.from(xml.querySelectorAll('item'))
-
-    const results = items
-      .map(item => ({
-        title: item.querySelector('title')?.textContent ?? 'No title',
-        magnet: getMagnetLink(item),
-        seeders: getAttributeValue(item, 'seeders'),
-        size: formatFileSize(item.querySelector('size')?.textContent ?? ''),
-      }))
-      .filter(item => item.magnet.startsWith('magnet:'))
-      .sort((a, b) => parseInt(b.seeders ?? '0') - parseInt(a.seeders ?? '0'))
-      .slice(0, 8)
-
-    console.log('Torrent search results:', results)
-
-    return { success: true, data: results }
-  } catch (error: any) {
-    return { success: false, error: error.message || 'Failed to parse XML' }
-  }
-}
-
-interface CalendarEventResource {
-  summary?: string
-  description?: string
-  start?: { dateTime?: string; timeZone?: string; date?: string }
-  end?: { dateTime?: string; timeZone?: string; date?: string }
-  location?: string
-  attendees?: { email: string }[]
-}
-
-/**
- * Retrieves a list of events from the user's Google Calendar.
- */
-
-async function get_calendar_events(args: {
-  calendarId?: string
-  timeMin?: string
-  timeMax?: string
-  q?: string
-  maxResults?: number
-}): Promise<FunctionResult> {
-  try {
-    const result = await window.ipcRenderer.invoke(
-      'google-calendar:list-events',
-      {
-        calendarId: args.calendarId || 'primary',
-        timeMin: args.timeMin,
-        timeMax: args.timeMax,
-        q: args.q,
-        maxResults: args.maxResults || 10,
-      }
-    )
-    if (result.success) {
-      return { success: true, data: result.data || 'No events found.' }
-    }
-    return {
-      success: false,
-      error: result.error || 'Failed to list calendar events.',
-    }
-  } catch (error: any) {
-    return { success: false, error: `IPC Error: ${error.message}` }
-  }
-}
-
-/**
- * Creates a new event in the user's Google Calendar.
- */
-
-async function create_calendar_event(args: {
-  calendarId?: string
-  summary: string
-  description?: string
-  startDateTime: string
-  endDateTime: string
-  location?: string
-  attendees?: string[]
-}): Promise<FunctionResult> {
-  try {
-    const eventResource: CalendarEventResource = {
-      summary: args.summary,
-      description: args.description,
-      start: { dateTime: args.startDateTime },
-      end: { dateTime: args.endDateTime },
-      location: args.location,
-    }
-    if (args.attendees && args.attendees.length > 0) {
-      eventResource.attendees = args.attendees.map(email => ({ email }))
-    }
-
-    const result = await window.ipcRenderer.invoke(
-      'google-calendar:create-event',
-      {
-        calendarId: args.calendarId || 'primary',
-        eventResource,
-      }
-    )
-    if (result.success) {
-      return { success: true, data: result.data }
-    }
-    return {
-      success: false,
-      error: result.error || 'Failed to create calendar event.',
-    }
-  } catch (error: any) {
-    return { success: false, error: `IPC Error: ${error.message}` }
-  }
-}
-
-/**
- * Updates an existing event in the user's Google Calendar.
- */
-
-async function update_calendar_event(args: {
-  calendarId?: string
-  eventId: string
-  summary?: string
-  description?: string
-  startDateTime?: string
-  endDateTime?: string
-  location?: string
-  attendees?: string[]
-}): Promise<FunctionResult> {
-  try {
-    const eventResource: CalendarEventResource = {}
-    if (args.summary) eventResource.summary = args.summary
-    if (args.description) eventResource.description = args.description
-    if (args.startDateTime)
-      eventResource.start = { dateTime: args.startDateTime }
-    if (args.endDateTime) eventResource.end = { dateTime: args.endDateTime }
-    if (args.location) eventResource.location = args.location
-    if (args.attendees && args.attendees.length > 0) {
-      eventResource.attendees = args.attendees.map(email => ({ email }))
-    }
-
-    if (Object.keys(eventResource).length === 0) {
-      return {
-        success: false,
-        error: 'No fields provided to update for the event.',
-      }
-    }
-
-    const result = await window.ipcRenderer.invoke(
-      'google-calendar:update-event',
-      {
-        calendarId: args.calendarId || 'primary',
-        eventId: args.eventId,
-        eventResource,
-      }
-    )
-    if (result.success) {
-      return { success: true, data: result.data }
-    }
-    return {
-      success: false,
-      error: result.error || 'Failed to update calendar event.',
-    }
-  } catch (error: any) {
-    return { success: false, error: `IPC Error: ${error.message}` }
-  }
-}
-
-/**
- * Deletes an event from the user's Google Calendar.
- */
-
-async function delete_calendar_event(args: {
-  calendarId?: string
-  eventId: string
-}): Promise<FunctionResult> {
-  try {
-    const result = await window.ipcRenderer.invoke(
-      'google-calendar:delete-event',
-      {
-        calendarId: args.calendarId || 'primary',
-        eventId: args.eventId,
-      }
-    )
-    if (result.success) {
-      return { success: true, data: result.data }
-    }
-    return {
-      success: false,
-      error: result.error || 'Failed to delete calendar event.',
-    }
-  } catch (error: any) {
-    return { success: false, error: `IPC Error: ${error.message}` }
-  }
-}
-
-/**
  * Helper function to process email details for AI
  */
 
@@ -744,298 +411,6 @@ async function get_email_content(
   }
 }
 
-async function list_directory(
-  args: ListDirectoryArgs
-): Promise<FunctionResult> {
-  try {
-    const result = await window.ipcRenderer.invoke(
-      'desktop:listDirectory',
-      args.path
-    )
-    if (result.success) {
-      return { success: true, data: result.files }
-    } else {
-      return { success: false, error: result.error }
-    }
-  } catch (error: any) {
-    return { success: false, error: error.message }
-  }
-}
-
-async function execute_command(
-  args: ExecuteCommandArgs
-): Promise<FunctionResult> {
-  try {
-    const settingsStore = useSettingsStore()
-    const commandName = args.command.split(' ')[0]
-
-    if (!settingsStore.isCommandApproved(args.command)) {
-      const approvalResult = await (window as any).requestCommandApproval(
-        args.command
-      )
-
-      if (!approvalResult.approved) {
-        return { success: false, error: 'Command execution denied by user' }
-      }
-    }
-
-    const result = await window.ipcRenderer.invoke(
-      'desktop:executeCommand',
-      args.command
-    )
-    if (result.success) {
-      return { success: true, data: result.output }
-    } else {
-      return { success: false, error: result.error }
-    }
-  } catch (error: any) {
-    return { success: false, error: error.message }
-  }
-}
-
-async function schedule_task(args: ScheduleTaskArgs): Promise<FunctionResult> {
-  try {
-    let cronExpression = parseNaturalLanguageToCron(args.schedule)
-
-    if (!cronExpression) {
-      if (validateCronExpression(args.schedule)) {
-        cronExpression = args.schedule
-      } else {
-        return {
-          success: false,
-          error: `Unable to parse schedule "${args.schedule}". Try formats like "every morning at 8 AM", "every hour", "daily at 6 PM", or use cron format like "0 8 * * *".`,
-        }
-      }
-    }
-
-    if (!validateCronExpression(cronExpression)) {
-      return {
-        success: false,
-        error: `Generated cron expression "${cronExpression}" is invalid.`,
-      }
-    }
-
-    const result = await window.ipcRenderer.invoke('scheduler:create-task', {
-      name: args.name,
-      cronExpression,
-      actionType: args.action_type,
-      details: args.details,
-    })
-
-    if (result.success) {
-      return {
-        success: true,
-        data: {
-          message: `Task "${args.name}" scheduled successfully.`,
-          taskId: result.taskId,
-          cronExpression,
-          schedule: args.schedule,
-        },
-      }
-    } else {
-      return {
-        success: false,
-        error: result.error || 'Failed to create scheduled task.',
-      }
-    }
-  } catch (error: any) {
-    return { success: false, error: error.message }
-  }
-}
-
-async function manage_scheduled_tasks(
-  args: ManageScheduledTasksArgs
-): Promise<FunctionResult> {
-  try {
-    switch (args.action) {
-      case 'list': {
-        const result = await window.ipcRenderer.invoke(
-          'scheduler:get-all-tasks'
-        )
-        if (result.success) {
-          const tasks = result.tasks.map((task: any) => ({
-            id: task.id,
-            name: task.name,
-            schedule: task.cronExpression,
-            actionType: task.actionType,
-            details: task.details,
-            isActive: task.isActive,
-            createdAt: task.createdAt,
-            lastRun: task.lastRun,
-            nextRun: task.nextRun,
-          }))
-          return {
-            success: true,
-            data: {
-              message: `Found ${tasks.length} scheduled tasks.`,
-              tasks,
-            },
-          }
-        } else {
-          return {
-            success: false,
-            error: result.error || 'Failed to get scheduled tasks.',
-          }
-        }
-      }
-
-      case 'delete': {
-        if (!args.task_id) {
-          return {
-            success: false,
-            error: 'Task ID is required for delete action.',
-          }
-        }
-
-        const result = await window.ipcRenderer.invoke(
-          'scheduler:delete-task',
-          {
-            taskId: args.task_id,
-          }
-        )
-
-        if (result.success) {
-          return {
-            success: true,
-            data: { message: `Task ${args.task_id} deleted successfully.` },
-          }
-        } else {
-          return {
-            success: false,
-            error: result.error || 'Failed to delete task.',
-          }
-        }
-      }
-
-      case 'toggle': {
-        if (!args.task_id) {
-          return {
-            success: false,
-            error: 'Task ID is required for toggle action.',
-          }
-        }
-
-        const result = await window.ipcRenderer.invoke(
-          'scheduler:toggle-task',
-          {
-            taskId: args.task_id,
-          }
-        )
-
-        if (result.success) {
-          return {
-            success: true,
-            data: {
-              message: `Task ${args.task_id} status toggled successfully.`,
-            },
-          }
-        } else {
-          return {
-            success: false,
-            error: result.error || 'Failed to toggle task status.',
-          }
-        }
-      }
-
-      default:
-        return { success: false, error: `Unknown action: ${args.action}` }
-    }
-  } catch (error: any) {
-    return { success: false, error: error.message }
-  }
-}
-
-/**
- * Helpers for parsing torrent data.
- */
-
-function formatFileSize(bytes: string): string {
-  const b = parseInt(bytes, 10)
-  if (isNaN(b)) return 'unknown'
-  const kb = b / 1024
-  const mb = kb / 1024
-  const gb = mb / 1024
-  if (gb > 1) return `${gb.toFixed(1)} GB`
-  if (mb > 1) return `${mb.toFixed(1)} MB`
-  return `${kb.toFixed(1)} KB`
-}
-
-function getAttributeValue(item: Element, attrName: string): string {
-  const attrs = item.getElementsByTagName('torznab:attr')
-  for (const attr of Array.from(attrs)) {
-    if (attr.getAttribute('name') === attrName) {
-      return attr.getAttribute('value') ?? ''
-    }
-  }
-  return ''
-}
-
-function getMagnetLink(item: Element): string {
-  const guid = item.querySelector('guid')?.textContent
-  if (guid?.startsWith('magnet:')) return guid
-
-  const attrs = item.getElementsByTagName('torznab:attr')
-  for (const attr of Array.from(attrs)) {
-    if (attr.getAttribute('name') === 'magneturl') {
-      return attr.getAttribute('value') ?? ''
-    }
-  }
-
-  return ''
-}
-
-/**
- * Adds a torrent to qBittorrent using the provided magnet link.
- */
-async function add_torrent_to_qb(
-  args: AddTorrentArgs
-): Promise<FunctionResult> {
-  const settings = useSettingsStore().config
-  const isDev = import.meta.env.DEV
-  const QB_BASE_URL = isDev ? '' : settings.VITE_QB_URL
-  const QB_USERNAME = settings.VITE_QB_USERNAME
-  const QB_PASSWORD = settings.VITE_QB_PASSWORD
-
-  if (!QB_USERNAME || !QB_PASSWORD) {
-    return { success: false, error: 'qBittorrent credentials not configured.' }
-  }
-  if (!args.magnet) {
-    return { success: false, error: 'Magnet link is missing.' }
-  }
-
-  console.log('Adding torrent to qBittorrent:', args.magnet)
-
-  try {
-    const loginRes = await axios.post(
-      `${QB_BASE_URL}/api/v2/auth/login`,
-      new URLSearchParams({ username: QB_USERNAME, password: QB_PASSWORD }),
-      { withCredentials: true }
-    )
-
-    if (loginRes.data !== 'Ok.') {
-      return { success: false, error: 'Login failed' }
-    }
-
-    await axios.post(
-      `${QB_BASE_URL}/api/v2/torrents/add`,
-      new URLSearchParams({ urls: args.magnet }),
-      {
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
-        withCredentials: true,
-      }
-    )
-
-    return { success: true, data: 'Torrent added successfully.' }
-  } catch (err: any) {
-    return { success: false, error: err.message || 'Failed to add torrent.' }
-  }
-}
-
-/**
- * Retrieves the current browser context, including URL, title, content, and selected text.
- */
 async function browser_context(
   args: BrowserContextArgs = {}
 ): Promise<FunctionResult> {
